Penyempurnaan Dokumen, Tim dan Dashboard

+++ la1/app/Http/Controllers/LA/TeamsController.php
@@ -23,7 +23,7 @@ class TeamsController extends Controller
 {
 	public $show_action = true;
 	public $view_col = 'nama';
-	public $listing_cols = ['id', 'nama', 'project_id', 'peran', 'nama_pj'];
+	public $listing_cols = ['id', 'nama', 'project_id', 'peran', 'nama_pj','staf','staf1','staf2','staf3'];
 	
 	public function __construct() {
 		// Field Access of Listing Columns
@@ -231,11 +231,11 @@ class TeamsController extends Controller
 			
 			if($this->show_action) {
 				$output = '';
-				if(Module::hasAccess("Teams", "edit")) {
+				if(Module::hasAccess("Teams", "edit")&&$data->data[$i][4]==Auth::user()->name) {
 					$output .= '<a href="'.url(config('laraadmin.adminRoute') . '/teams/'.$data->data[$i][0].'/edit').'" class="btn btn-warning btn-xs" style="display:inline;padding:2px 5px 3px 5px;"><i class="fa fa-edit"></i></a>';
 				}
 				
-				if(Module::hasAccess("Teams", "delete")) {
+				if(Module::hasAccess("Teams", "delete")&&$data->data[$i][4]==Auth::user()->name) {
 					$output .= Form::open(['route' => [config('laraadmin.adminRoute') . '.teams.destroy', $data->data[$i][0]], 'method' => 'delete', 'style'=>'display:inline']);
 					$output .= ' <button class="btn btn-danger btn-xs" type="submit"><i class="fa fa-times"></i></button>';
 					$output .= Form::close();
